Moved more public key tests to the public_keys test module
